example_pillow/pyproject.toml
[tool.poetry]
name = "example_pillow"
version = "0.1.0"
description = ""
authors = ["Theo Bob Massard <tbobm@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
